refactor(state): split inventoryStore.ts into focused modules
inventoryStore.ts had grown to 1,115 lines, the largest store in the
codebase. Splits it into a thin store entry point plus sibling modules,
mirroring the loreStore/narrativeStore decomposition already in
src/state/:

- inventoryStore.types.ts: store interface + payload types
- inventoryStore.state.ts: initial state
- inventoryStore.helpers.ts: sanitization, validation, journal-entry
  creation, and the shared ensureCharacterInventory factory
- inventoryStore.actions.ts: core item CRUD + image-generation tracking
- inventoryStore.characterInventory.ts: add/remove/query a character's
  items, including stacking
- inventoryStore.equipment.ts: item usage + equip/unequip
- inventoryStore.persistence.ts: IndexedDB persist config
- inventoryStore.ts: composes the above, plus store-event subscriptions

Pure refactor — no behavior change. Public API (hook, exported types)
is unchanged.

Fixes #1415
